stable-diff/requirements.txt
jupyterlab==3.2.9
nbformat==5.4.0
notebook==6.4.10
ftfy==6.1.1
matplotlib==3.5.3
torch>=2.0.1
omegaconf>=2.3.0
pandas>=2.0.3
pillow>=9.5.0
diffusers==0.23.1
transformers==4.35.2
tokenizers==0.15.0
safetensors==0.4.0
scipy==1.11.4
einops==0.7.0
fire==0.5.0
opencv-python==4.8.1.78
accelerate==0.24.1
clip @ git+https://github.com/openai/CLIP.git@a1d071733d7111c9c014f024669f959182114e33
metaflow-checkpoint